+++ common/src/main/java/smartin/miapi/client/atlas/MaterialSpriteManager.java
@@ -15,13 +15,13 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.entity.ItemRenderer;
 import net.minecraft.client.renderer.texture.DynamicTexture;
-import net.minecraft.client.renderer.texture.SpriteContents;
 import net.minecraft.client.renderer.texture.TextureAtlas;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import smartin.miapi.Miapi;
 import smartin.miapi.client.AnimatedTexturesManager;
 import smartin.miapi.client.MiapiClient;
 import smartin.miapi.client.renderer.MovedVertexConsumer;
@@ -48,8 +48,6 @@ import java.util.function.IntUnaryOperator;
 @Environment(EnvType.CLIENT)
 public class MaterialSpriteManager {
     static Map<Holder, DynamicTexture> animated_Textures = new HashMap<>();
-    static long CURRENT_GENERATION = 0;
-
     public static final long CACHE_SIZE = 10000;
     public static final long CACHE_LIFETIME = 10;
     public static final TimeUnit CACHE_LIFETIME_UNIT = TimeUnit.SECONDS;
@@ -106,6 +104,9 @@ public class MaterialSpriteManager {
         });
     }
 
+    /**
+     * creates or finds an existing off atlas Dynamic Texture Sprite and returns the id
+     */
     public static ResourceLocation getMaterialSprite(Holder holder) {
         ResourceLocation identifier = materialSpriteCache.getIfPresent(holder);
         if (identifier == null) {
@@ -131,58 +132,71 @@ public class MaterialSpriteManager {
         materialSpriteCache.invalidateAll();
         ANIMATED_ATLAS_SPRITES.clear();
         FAST_CACHE.forEach((h, slot) -> {
-            slot.used = 0;
+            slot.invalidate();
             ATLAS_SPRITE_POOL.computeIfAbsent(resToKey(slot.x, slot.y), (v) -> new ArrayList<>()).add(slot);
         });
         FAST_CACHE.clear();
-        //TODO:free atlas sprites
+        nativeImageBackedTextureMap.forEach((id, texture) -> {
+            texture.close();
+        });
+        nativeImageBackedTextureMap.clear();
+        Miapi.LOGGER.info("clear at tick " + MiapiClient.tick.get());
     }
 
     /**
      * the tick function.
      * responsible for animated textures and clearing unused caches
+     * also responsible for managing SpriteSlot activity and decay
      */
     public static void tick() {
         if (ReloadEvents.isInReload()) {
             return;
         }
-        List<Holder> toRemove = new ArrayList<>();
-        animated_Textures.forEach(((holder, nativeImageBackedTexture) -> {
-            try {
-                holder.colorer.tick((nativeImage) -> {
-                    //important!
-                    //the MaskColorer is responsible for managing any NativeImage it creates.
-                    //BUT the NativeBackedTexture removes its old uploaded NativeImage, so we need to upload a copy
-                    Objects.requireNonNull(nativeImageBackedTexture.getPixels()).copyFrom(nativeImage);
-                    nativeImageBackedTexture.upload();
-                }, holder.sprite().contents());
-            } catch (Exception e) {
-                toRemove.add(holder);
+        try {
+            List<Holder> toRemove = new ArrayList<>();
+            animated_Textures.forEach(((holder, nativeImageBackedTexture) -> {
+                try {
+                    holder.colorer.tick((nativeImage) -> {
+                        //important!
+                        //the MaskColorer is responsible for managing any NativeImage it creates.
+                        //BUT the NativeBackedTexture removes its old uploaded NativeImage, so we need to upload a copy
+                        Objects.requireNonNull(nativeImageBackedTexture.getPixels()).copyFrom(nativeImage);
+                        nativeImageBackedTexture.upload();
+                    }, holder.sprite().contents());
+                } catch (Exception e) {
+                    toRemove.add(holder);
+                }
+            }));
+            ANIMATED_ATLAS_SPRITES.forEach(slot -> {
+                AnimatedTexturesManager.markAnimated(slot.getSprite());
+                slot.updateSprite();
+            });
+            toRemove.forEach(materialSpriteCache::invalidate);
+            List<Holder> toRemoveFAST = new ArrayList<>();
+            FAST_CACHE.forEach((h, s) -> {
+                s.tickUsage();
+                if (s.shouldBeFreed()) {
+                    toRemoveFAST.add(h);
+                    s.clear();
+                    ANIMATED_ATLAS_SPRITES.remove(s);
+                }
+            });
+            toRemoveFAST.forEach(h -> {
+                SpriteSlot slot = FAST_CACHE.remove(h);
+                ATLAS_SPRITE_POOL.computeIfAbsent(resToKey(slot.x, slot.y), (s) -> new ArrayList<>()).add(slot);
+            });
+            for (TextureAtlasSprite sprite : animated) {
+                AnimatedTexturesManager.markAnimated(sprite);
             }
-        }));
-        ANIMATED_ATLAS_SPRITES.forEach(slot -> {
-            AnimatedTexturesManager.markAnimated(slot.getSprite());
-            slot.updateSprite();
-        });
-        toRemove.forEach(materialSpriteCache::invalidate);
-        List<Holder> toRemoveFAST = new ArrayList<>();
-        FAST_CACHE.forEach((h, s) -> {
-            s.used--;
-            if (s.used < 1) {
-                toRemoveFAST.add(h);
-                s.clear();
-                ANIMATED_ATLAS_SPRITES.remove(s);
-            }
-        });
-        toRemoveFAST.forEach(h -> {
-            SpriteSlot slot = FAST_CACHE.remove(h);
-            ATLAS_SPRITE_POOL.computeIfAbsent(resToKey(slot.x, slot.y), (s) -> new ArrayList<>()).add(slot);
-        });
-        for (TextureAtlasSprite sprite : animated) {
-            AnimatedTexturesManager.markAnimated(sprite);
+        } catch (RuntimeException e) {
+            Miapi.LOGGER.info("prevented crash at tick " + MiapiClient.tick.get());
+            MiapiEvents.CLEAR_CACHE.invoker().onReload();
         }
     }
 
+    /**
+     * redo the setup for the Provider, grabing (if available) a space on the BlockAtlas
+     */
     public static void providerReAquire(VertexConsumerProvider provider) {
         Holder holder = provider.spriteHolder;
         provider.isFast = false;
@@ -191,7 +205,9 @@ public class MaterialSpriteManager {
             if (spriteSlot == null) {
                 spriteSlot = getFreeAtlasSlot(((SpriteContentsAccessor) holder.sprite().contents()).getMiapiWidth(), ((SpriteContentsAccessor) holder.sprite().contents()).getMiapiHeight());
                 if (spriteSlot != null) {
-                    spriteSlot.used = 6;
+                    //set to 4 to block rendering in first frame to to upload preceding rendering,
+                    //so a SpriteSlot is not available during first requested frame
+                    spriteSlot.allocated();
                     spriteSlot.holder = provider.spriteHolder;
                     FAST_CACHE.put(provider.spriteHolder, spriteSlot);
                     if (provider.spriteHolder.colorer().doTick()) {
@@ -203,9 +219,15 @@ public class MaterialSpriteManager {
             }
             if (spriteSlot != null) {
                 provider.spriteSlot = spriteSlot;
-                spriteSlot.clearOwner.add(() -> provider.spriteSlot = null);
             }
         }
+    }
+
+
+    public static boolean isSlotStilLValid(VertexConsumerProvider provider) {
+        return provider.spriteSlot != null &&
+               provider.spriteSlot.canBeRendered() &&
+               provider.spriteSlot.holder != null && provider.spriteSlot.holder.hashCode() == provider.spriteHolder.hashCode();
     }
 
     public static void markTextureAsAnimatedInUse(TextureAtlasSprite sprite) {
@@ -242,14 +264,12 @@ public class MaterialSpriteManager {
             return;
         }
         out.getRenderSaveVC = (b -> {
-            if (out.spriteSlot == null) {
-                providerReAquire(out);
-            } else if (out.spriteSlot.used > 0 && out.spriteSlot.used < 4 && out.spriteSlot.holder != null && out.spriteSlot.holder.hashCode() == out.spriteHolder.hashCode()) {
+            if (isSlotStilLValid(out)) {
                 return getBlockAtlasVertexConsumer(b, originalSprite, out.spriteHolder, out.spriteSlot);
             } else {
-                out.spriteSlot.clearOwner();
+                providerReAquire(out);
+                return getDynamicTextureVertexConsumer(b, originalSprite, out.spriteHolder);
             }
-            return getDynamicTextureVertexConsumer(b, originalSprite, out.spriteHolder);
         });
         //add sprite verification checks once per tick to each out vcprovider
         out.vanillaVCGetter = MaterialSpriteManager::getVanillaItemVC;
@@ -259,8 +279,8 @@ public class MaterialSpriteManager {
      * Creates a Vertex Consumer that uses on BlockAtlas Textures to render (if possible)
      */
     private static @NotNull VertexConsumer getBlockAtlasVertexConsumer(MultiBufferSource vertexConsumers, TextureAtlasSprite originalSprite, Holder holder, SpriteSlot spriteSlot) {
-        if (spriteSlot.used < 3) {
-            spriteSlot.used = 3;
+        if (spriteSlot.canBeRendered()) {
+            spriteSlot.use();
         }
         return new MovedVertexConsumer(
                 getVanillaItemVC(vertexConsumers),
@@ -279,6 +299,11 @@ public class MaterialSpriteManager {
         return new RescaledVertexConsumer(atlasConsumer, originalSprite);
     }
 
+    /**
+     * Holders serve as unique Identifiers for any kind of texture by how its collored.
+     * (Prob) this should not require the material and just the SpriteColorer in the future.
+     * is extremly hash and equal friendly.
+     */
     public record Holder(TextureAtlasSprite sprite, Material material, SpriteColorer colorer, int hash) {
 
         public Holder(TextureAtlasSprite sprite, Material material, SpriteColorer colorer) {
@@ -290,15 +315,14 @@ public class MaterialSpriteManager {
             if (this == o) return true;
             if (o == null || getClass() != o.getClass()) return false;
             Holder holder = (Holder) o;
-            return Objects.equals(sprite, holder.sprite) &&
+            return this.hash() == holder.hash() &&
+                   Objects.equals(sprite, holder.sprite) &&
                    Objects.equals(material, holder.material) &&
                    Objects.equals(colorer, holder.colorer);
         }
 
         @Override
-        public int hashCode() {
-            return hash;
-        }
+        public int hashCode() {return hash;}
     }
 
     /**
@@ -316,9 +340,8 @@ public class MaterialSpriteManager {
 
         for (int i = 0; i < slots.size(); i++) {
             SpriteSlot slot = slots.get(i);
-            if (slot.used < 1) {
+            if (slot.shouldBeFreed()) {
                 slots.remove(slot);
-
                 return slot;
             }
         }
@@ -331,26 +354,24 @@ public class MaterialSpriteManager {
      * @param width  the desired sprite width
      * @param height the desired sprite height
      */
-    public static int resToKey(int width, int height) {
-        return (width << 16) | height;
-    }
+    public static int resToKey(int width, int height) {return (width << 16) | height;}
 
     /**
      * a slot on the main block atlas for a sprite, manages its content and stuffs.
      */
     public static class SpriteSlot {
-        public int used = 0;
-        public long generation = 0;
+        private int used = 0;
         public ResourceLocation internalID;
-        public Consumer<NativeImage> update;
+        private Consumer<NativeImage> update;
         public int x;
         public int y;
-        public final SpriteContents contents;
+        public final BufferSpriteAdder.MiapiSpriteContents contents;
         public TextureAtlasSprite sprite;
         public Holder holder;
-        public List<Runnable> clearOwner = new ArrayList<>();
+        private static final int KEEP_ALIVE_COUNT = 10;
+        private static final int MAX_RENDER_ALLOWED = KEEP_ALIVE_COUNT - 1;
 
-        public SpriteSlot(ResourceLocation id, int x, int y, SpriteContents contents, Consumer<NativeImage> update) {
+        public SpriteSlot(ResourceLocation id, int x, int y, BufferSpriteAdder.MiapiSpriteContents contents, Consumer<NativeImage> update) {
             this.contents = contents;
             this.internalID = id;
             this.x = x;
@@ -365,6 +386,20 @@ public class MaterialSpriteManager {
             return sprite;
         }
 
+        public boolean canBeRendered() {return used <= MAX_RENDER_ALLOWED && used > 0 && isUploaded();}
+
+        public boolean shouldBeFreed() {return used < 1;}
+
+        public void allocated() {used = KEEP_ALIVE_COUNT;}
+
+        public boolean isUploaded() {return !contents.dirty;}
+
+        public void use() {used = MAX_RENDER_ALLOWED;}
+
+        public void tickUsage() {used--;}
+
+        public void invalidate() {used = 0;}
+
         public int width() {
             return x;
         }
@@ -377,11 +412,6 @@ public class MaterialSpriteManager {
             if (holder != null) {
                 update.accept(holder.colorer().createSpriteManager(holder.sprite().contents()).recolor());
             }
-            this.generation = CURRENT_GENERATION;
-        }
-
-        public void release() {
-            clearOwner();
         }
 
         public void clear() {
@@ -392,16 +422,10 @@ public class MaterialSpriteManager {
                 }
             }
             update.accept(image);
-            clearOwner();
-        }
-
-        public void clearOwner() {
-            clearOwner.forEach(Runnable::run);
-            clearOwner.clear();
+            used = 0;
         }
 
         public void destroy() {
-            clearOwner();
             used = 0;
             if (contents != null) {
                 contents.close();