+++ common/src/main/java/smartin/miapi/mixin/client/NativeImageAccessor.java
@@ -1,0 +1,11 @@
+package smartin.miapi.mixin.client;
+
+import com.mojang.blaze3d.platform.NativeImage;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(NativeImage.class)
+public interface NativeImageAccessor {
+    @Accessor("pixels")
+    long getPixelsMiapi();
+}